refactoring GetPublicData
